move CardGroup inside DisplayCard grouping and reorganize sx
