--- @ndlib/gatsby-theme-marble/src/components/Shared/CardGroup/CardGroupToggle/sx.js
@@ -1,11 +1,0 @@
-module.exports = {
-  wrapper: {
-    justifyContent: 'space-between',
-    '& input': {
-      borderRadius: '.25rem',
-      margin: '2px',
-      padding: '2px',
-      width : '30px',
-    },
-  },
-}
