+++ @ndlib/gatsby-theme-marble/src/components/Shared/SiblingItems/index.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import Link from 'components/Shared/Link'
 import { jsx, Heading } from 'theme-ui'
 import typy from 'typy'
-import CardGroup from 'components/Shared/CardGroup'
+import CardGroup from 'components/Shared/DisplayCard/CardGroup'
 import MarbleItemCard from 'components/Shared/DisplayCard/MarbleItemCard'
 import getArrayNeighbors from 'utils/getArrayNeighbors'
 import findImage from 'utils/findImage'
